代码转换_3/paddle_project/requirements.txt
tqdm
numpy
paddlepaddle-gpu==1.0.1
pyyaml
librosa
mir_eval
matplotlib
tensorboardX
Pillow
